refactor: Centralize default models and providers in constants
