+++ src/components/settings/EmbeddingDbManageRoot.tsx
@@ -5,7 +5,6 @@ import { Loader2, PickaxeIcon, RefreshCw, Trash2 } from 'lucide-react'
 import { Notice } from 'obsidian'
 import { useState } from 'react'
 
-import { EMBEDDING_MODEL_OPTIONS } from '../../constants'
 import { useDatabase } from '../../contexts/database-context'
 import { useSettings } from '../../contexts/settings-context'
 import { getEmbeddingModelClient } from '../../core/rag/embedding'
@@ -36,10 +35,10 @@ export default function EmbeddingDbManageRoot({
 
       const statsMap = new Map(dbStats.map((stat) => [stat.model, stat]))
 
-      return EMBEDDING_MODEL_OPTIONS.map((option) => ({
-        model: option.id,
-        rowCount: statsMap.get(option.id)?.rowCount ?? 0,
-        totalDataBytes: statsMap.get(option.id)?.totalDataBytes ?? 0,
+      return settings.embeddingModels.map((embeddingModel) => ({
+        model: embeddingModel.id,
+        rowCount: statsMap.get(embeddingModel.id)?.rowCount ?? 0,
+        totalDataBytes: statsMap.get(embeddingModel.id)?.totalDataBytes ?? 0,
       }))
     },
   })
