+++ src/constants.ts
@@ -1,235 +1,9 @@
-import { EmbeddingModelOption } from './types/embedding'
-import { ModelOption } from './types/llm/model'
+import { ChatModel } from './types/chat-model.types'
+import { EmbeddingModel } from './types/embedding-model.types'
+import { LLMProvider, LLMProviderType } from './types/provider.types'
 
 export const CHAT_VIEW_TYPE = 'smtcmp-chat-view'
 export const APPLY_VIEW_TYPE = 'smtcmp-apply-view'
-
-// TODO: remove this
-export const CHAT_MODEL_OPTIONS: ModelOption[] = [
-  {
-    id: 'anthropic/claude-3.5-sonnet-latest',
-    name: 'claude-3.5-sonnet (Recommended)',
-    model: {
-      provider: 'anthropic',
-      model: 'claude-3-5-sonnet-latest',
-    },
-  },
-  {
-    id: 'openai/gpt-4o',
-    name: 'gpt-4o',
-    model: {
-      provider: 'openai',
-      model: 'gpt-4o',
-    },
-  },
-  {
-    id: 'openai/gpt-4o-mini',
-    name: 'gpt-4o-mini',
-    model: {
-      provider: 'openai',
-      model: 'gpt-4o-mini',
-    },
-  },
-  {
-    id: 'gemini/gemini-1.5-pro',
-    name: 'gemini-1.5-pro',
-    model: {
-      provider: 'gemini',
-      model: 'gemini-1.5-pro',
-    },
-  },
-  {
-    id: 'gemini/gemini-2.0-flash',
-    name: 'gemini-2.0-flash',
-    model: {
-      provider: 'gemini',
-      model: 'gemini-2.0-flash-exp',
-    },
-  },
-  {
-    id: 'gemini/gemini-2.0-flash-thinking',
-    name: 'gemini-2.0-flash-thinking',
-    model: {
-      provider: 'gemini',
-      model: 'gemini-2.0-flash-thinking-exp',
-    },
-  },
-  {
-    id: 'gemini/gemini-exp-1206',
-    name: 'gemini-exp-1206',
-    model: {
-      provider: 'gemini',
-      model: 'gemini-exp-1206',
-    },
-  },
-  {
-    id: 'groq/llama-3.1-70b-versatile',
-    name: 'llama-3.1-70b (Groq)',
-    model: {
-      provider: 'groq',
-      model: 'llama-3.1-70b-versatile',
-    },
-  },
-  {
-    id: 'ollama',
-    name: 'Ollama',
-    model: {
-      provider: 'ollama',
-      model: '',
-      baseURL: '',
-    },
-  },
-  {
-    id: 'openai-compatible',
-    name: 'Custom (OpenAI Compatible)',
-    model: {
-      provider: 'openai-compatible',
-      model: '',
-      apiKey: '',
-      baseURL: '',
-    },
-  },
-  {
-    id: 'openai/o1',
-    name: 'OpenAI o1 (Non-Streaming)',
-    model: {
-      provider: 'openai',
-      model: 'o1',
-      supportsStreaming: false,
-    },
-  },
-]
-
-export const APPLY_MODEL_OPTIONS: ModelOption[] = [
-  {
-    id: 'openai/gpt-4o-mini',
-    name: 'gpt-4o-mini (Recommended)',
-    model: {
-      provider: 'openai',
-      model: 'gpt-4o-mini',
-    },
-  },
-  {
-    id: 'anthropic/claude-3.5-haiku',
-    name: 'claude-3.5-haiku',
-    model: {
-      provider: 'anthropic',
-      model: 'claude-3-5-haiku-latest',
-    },
-  },
-  {
-    id: 'gemini/gemini-1.5-flash',
-    name: 'gemini-1.5-flash',
-    model: {
-      provider: 'gemini',
-      model: 'gemini-1.5-flash',
-    },
-  },
-  {
-    id: 'gemini/gemini-2.0-flash',
-    name: 'gemini-2.0-flash',
-    model: {
-      provider: 'gemini',
-      model: 'gemini-2.0-flash-exp',
-    },
-  },
-  {
-    id: 'groq/llama-3.1-8b-instant',
-    name: 'llama-3.1-8b (Groq)',
-    model: {
-      provider: 'groq',
-      model: 'llama-3.1-8b-instant',
-    },
-  },
-  {
-    id: 'groq/llama-3.1-70b-versatile',
-    name: 'llama-3.1-70b (Groq)',
-    model: {
-      provider: 'groq',
-      model: 'llama-3.1-70b-versatile',
-    },
-  },
-  {
-    id: 'ollama',
-    name: 'Ollama',
-    model: {
-      provider: 'ollama',
-      model: '',
-      baseURL: '',
-    },
-  },
-  {
-    id: 'openai-compatible',
-    name: 'Custom (OpenAI Compatible)',
-    model: {
-      provider: 'openai-compatible',
-      model: '',
-      apiKey: '',
-      baseURL: '',
-    },
-  },
-]
-
-// Update table exports in src/database/schema.ts when updating this
-export const EMBEDDING_MODEL_OPTIONS: EmbeddingModelOption[] = [
-  {
-    id: 'openai/text-embedding-3-small',
-    name: 'openai/text-embedding-3-small (Recommended)',
-    model: {
-      provider: 'openai',
-      model: 'text-embedding-3-small',
-    },
-    dimension: 1536,
-  },
-  {
-    id: 'openai/text-embedding-3-large',
-    name: 'openai/text-embedding-3-large',
-    model: {
-      provider: 'openai',
-      model: 'text-embedding-3-large',
-    },
-    dimension: 3072,
-  },
-  {
-    id: 'gemini/text-embedding-004',
-    name: 'gemini/text-embedding-004',
-    model: {
-      provider: 'gemini',
-      model: 'text-embedding-004',
-    },
-    dimension: 768,
-  },
-  {
-    id: 'ollama/nomic-embed-text',
-    name: 'ollama/nomic-embed-text',
-    model: {
-      provider: 'ollama',
-      model: 'nomic-embed-text',
-      baseURL: '',
-    },
-    dimension: 768,
-  },
-  {
-    id: 'ollama/mxbai-embed-large',
-    name: 'ollama/mxbai-embed-large',
-    model: {
-      provider: 'ollama',
-      model: 'mxbai-embed-large',
-      baseURL: '',
-    },
-    dimension: 1024,
-  },
-  {
-    id: 'ollama/bge-m3',
-    name: 'ollama/bge-m3',
-    model: {
-      provider: 'ollama',
-      model: 'bge-m3',
-      baseURL: '',
-    },
-    dimension: 1024,
-  },
-]
 
 // Pricing in dollars per million tokens
 type ModelPricing = {
@@ -237,6 +11,7 @@ type ModelPricing = {
   output: number
 }
 
+// TODO: ensure we use model name (not our id) on checking price
 export const OPENAI_PRICES: Record<string, ModelPricing> = {
   'gpt-4o': { input: 2.5, output: 10 },
   'gpt-4o-mini': { input: 0.15, output: 0.6 },
@@ -259,3 +34,156 @@ export const GROQ_PRICES: Record<string, ModelPricing> = {
 }
 
 export const PGLITE_DB_PATH = '.smtcmp_vector_db.tar.gz'
+
+export const DEFAULT_PROVIDER_IDS = {
+  openai: 'openai',
+  anthropic: 'anthropic',
+  gemini: 'gemini',
+  groq: 'groq',
+  ollama: 'ollama',
+  'openai-compatible': null, // no default provider for this type
+} satisfies Record<LLMProviderType, string | null>
+
+export const DEFAULT_PROVIDERS: readonly LLMProvider[] = [
+  {
+    type: 'openai',
+    id: DEFAULT_PROVIDER_IDS.openai,
+  },
+  {
+    type: 'anthropic',
+    id: DEFAULT_PROVIDER_IDS.anthropic,
+  },
+  {
+    type: 'gemini',
+    id: DEFAULT_PROVIDER_IDS.gemini,
+  },
+  {
+    type: 'groq',
+    id: DEFAULT_PROVIDER_IDS.groq,
+  },
+  {
+    type: 'ollama',
+    id: DEFAULT_PROVIDER_IDS.ollama,
+  },
+]
+
+export const DEFAULT_CHAT_MODELS: readonly ChatModel[] = [
+  {
+    providerType: 'anthropic',
+    providerId: DEFAULT_PROVIDER_IDS.anthropic,
+    id: 'claude-3.5-sonnet',
+    model: 'claude-3-5-sonnet-latest',
+  },
+  {
+    providerType: 'openai',
+    providerId: DEFAULT_PROVIDER_IDS.openai,
+    id: 'gpt-4o',
+    model: 'gpt-4o',
+  },
+  {
+    providerType: 'openai',
+    providerId: DEFAULT_PROVIDER_IDS.openai,
+    id: 'gpt-4o-mini',
+    model: 'gpt-4o-mini',
+  },
+  {
+    providerType: 'gemini',
+    providerId: DEFAULT_PROVIDER_IDS.gemini,
+    id: 'gemini-1.5-pro',
+    model: 'gemini-1.5-pro',
+  },
+  {
+    providerType: 'gemini',
+    providerId: DEFAULT_PROVIDER_IDS.gemini,
+    id: 'gemini-2.0-flash',
+    model: 'gemini-2.0-flash-exp',
+  },
+  {
+    providerType: 'gemini',
+    providerId: DEFAULT_PROVIDER_IDS.gemini,
+    id: 'gemini-2.0-flash-thinking',
+    model: 'gemini-2.0-flash-thinking-exp',
+  },
+  {
+    providerType: 'gemini',
+    providerId: DEFAULT_PROVIDER_IDS.gemini,
+    id: 'gemini-exp-1206',
+    model: 'gemini-exp-1206',
+  },
+  {
+    providerType: 'groq',
+    providerId: DEFAULT_PROVIDER_IDS.groq,
+    id: 'groq/llama-3.1-70b',
+    model: 'llama-3.1-70b-versatile',
+  },
+  {
+    providerType: 'openai',
+    providerId: DEFAULT_PROVIDER_IDS.openai,
+    id: 'o1',
+    model: 'o1',
+    streamingDisabled: true, // currently, o1 API doesn't support streaming
+  },
+  {
+    providerType: 'anthropic',
+    providerId: DEFAULT_PROVIDER_IDS.anthropic,
+    id: 'claude-3.5-haiku',
+    model: 'claude-3-5-haiku-latest',
+  },
+  {
+    providerType: 'gemini',
+    providerId: DEFAULT_PROVIDER_IDS.gemini,
+    id: 'gemini-1.5-flash',
+    model: 'gemini-1.5-flash',
+  },
+  {
+    providerType: 'groq',
+    providerId: DEFAULT_PROVIDER_IDS.groq,
+    id: 'groq/llama-3.1-8b',
+    model: 'llama-3.1-8b-instant',
+  },
+]
+
+export const DEFAULT_EMBEDDING_MODELS: readonly EmbeddingModel[] = [
+  {
+    providerType: 'openai',
+    providerId: DEFAULT_PROVIDER_IDS.openai,
+    id: 'openai/text-embedding-3-small',
+    model: 'text-embedding-3-small',
+    dimension: 1536,
+  },
+  {
+    providerType: 'openai',
+    providerId: DEFAULT_PROVIDER_IDS.openai,
+    id: 'openai/text-embedding-3-large',
+    model: 'text-embedding-3-large',
+    dimension: 3072,
+  },
+  {
+    providerType: 'gemini',
+    providerId: DEFAULT_PROVIDER_IDS.gemini,
+    id: 'gemini/text-embedding-004',
+    model: 'text-embedding-004',
+    dimension: 768,
+  },
+  {
+    providerType: 'ollama',
+    providerId: DEFAULT_PROVIDER_IDS.ollama,
+    id: 'ollama/nomic-embed-text',
+    model: 'nomic-embed-text',
+    dimension: 768,
+  },
+  {
+    providerType: 'ollama',
+    providerId: DEFAULT_PROVIDER_IDS.ollama,
+    id: 'ollama/mxbai-embed-large',
+    model: 'mxbai-embed-large',
+    dimension: 1024,
+  },
+  {
+    providerType: 'ollama',
+    providerId: DEFAULT_PROVIDER_IDS.ollama,
+    id: 'ollama/bge-m3',
+    model: 'bge-m3',
+    dimension: 1024,
+  },
+]
