+++ src/settings/EmbeddingDbManageModal.tsx
@@ -2,11 +2,11 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { App, Modal } from 'obsidian'
 import React from 'react'
 import { createRoot } from 'react-dom/client'
-import { SettingsProvider } from 'src/contexts/settings-context'
 
 import EmbeddingDbManageRoot from '../components/settings/EmbeddingDbManageRoot'
 import { AppProvider } from '../contexts/app-context'
 import { DatabaseProvider } from '../contexts/database-context'
+import { SettingsProvider } from '../contexts/settings-context'
 import SmartCopilotPlugin from '../main'
 
 export class EmbeddingDbManageModal extends Modal {
