+++ src/utils/ollama.ts
@@ -1,5 +1,6 @@
 import { requestUrl } from 'obsidian'
 
+// TODO: remove this if not needed
 export async function getOllamaModels(ollamaUrl: string) {
   try {
     const response = (await requestUrl(`${ollamaUrl}/api/tags`)).json as {
